🎉 Clean start to the Open Source client
This is the latest copy of the Notarize Node/JS client with all the
tests and GitHub Workflow tests in-place but no history, and no other
branches. This will be a clean start for the use of the library by
others as well as being a destionation for new requests/PRs from others.