🔧 Upgrade loads of dependencies in clean-up
We had to really upgrade all the dependencies for the library as the
latest snakecase-keys and camelcase-keys weren't going to fly with the
old Typescript and their dependencies. Given that this is so focused a
library, and we know exactly what's being done, the impact of the
changes are really minor. So here we go.