🧪 Fix ESM Issues with camelcase-keys
In `camelcase-keys` 8.x or 9.x, we have an ESM `reuiqre()` problem, and
the solutions I was able to find are not at all simple. Sadly, this is
all about `ts-node`, and not the code itself. But we need that for the
tests. So... here we have several changes.

* we reverted to `camelcase-keys` `7.0.2` as that works.
* we removed all `require()` uses in the app, and this was nice, but it
  wasn't without it's own issues. Still, the new code is just as nice,
  and a wee bit cleaner.
* remove all unnecessary dependencies

All told, this seems to be working just fine. Tests and all.